parallel_for: fixed warning

+++ include/igl/parallel_for.h
@@ -130,7 +130,7 @@ inline bool igl::parallel_for(
 #ifdef IGL_PARALLEL_FOR_FORCE_SERIAL
     0;
 #else
-    loop_size<min_parallel?0:(sthc==0?8:sthc);
+    size_t(loop_size)<min_parallel?0:(sthc==0?8:sthc);
 #endif
   if(nthreads==0)
   {
